Refactoring: move setting of debuff potion to func
